A simple music app that passes CTS tests.
Deleted most of the code from the previous app. There is Music icon for this app
because CTS tests that tests MediaBrowserService and Playback callbacks.

It pops up a dialog when an audio file is clicked on and when it is playing (same as
before).

The following test is the only test that tests Music app.
  cts/tests/tests/content/src/android/content/cts/AvailableIntentsTest.java

Test: manual - Ran the following CTS tests on Pixel phone.
./development/testrunner/runtest.py --path cts/tests/tests/content/src/android/content/cts/AvailableIntentsTest.java
./development/testrunner/runtest.py --path cts/tests/tests/media/src/android/media/cts/MediaRouterTest.java
./development/testrunner/runtest.py --path cts/tests/tests/media/src/android/media/cts/AudioManagerTest.java
./development/testrunner/runtest.py --path cts/tests/tests/media/src/android/media/cts/MediaSessionTest.java
./development/testrunner/runtest.py --path cts/tests/tests/tv/src/android/media/tv/cts/TvContractTest.java

Bug: 112925066
Change-Id: I81080f138774c52ea636ddfca3eff26306f2f051